+++ android/app/src/main/java/co/edgesecure/app/MainActivity.kt
@@ -8,6 +8,7 @@ import com.facebook.react.ReactActivityDelegate
 import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint.fabricEnabled
 import com.facebook.react.defaults.DefaultReactActivityDelegate
 import com.zoontek.rnbootsplash.RNBootSplash
+import expo.modules.ReactActivityDelegateWrapper
 
 class MainActivity : ReactActivity() {
   /**
@@ -18,10 +19,16 @@ class MainActivity : ReactActivity() {
 
   /**
    * Returns the instance of the [ReactActivityDelegate]. We use [DefaultReactActivityDelegate]
-   * which allows you to enable New Architecture with a single boolean flags [fabricEnabled]
+   * which allows you to enable New Architecture with a single boolean flags [fabricEnabled].
+   * The Expo wrapper forwards activity lifecycle events (onCreate intent capture, onNewIntent)
+   * to expo modules; expo-quick-actions needs it to deliver shortcut taps on cold start.
    */
   override fun createReactActivityDelegate(): ReactActivityDelegate =
-    DefaultReactActivityDelegate(this, mainComponentName, fabricEnabled)
+    ReactActivityDelegateWrapper(
+      this,
+      BuildConfig.IS_NEW_ARCHITECTURE_ENABLED,
+      DefaultReactActivityDelegate(this, mainComponentName, fabricEnabled)
+    )
 
   // Edge addition
   override fun onCreate(savedInstanceState: Bundle?) {