+++ scripts/prepare.sh
@@ -11,7 +11,7 @@ cd "$(dirname "$0")/.."
 node -r sucrase/register ./scripts/configure.ts
 
 ## Fix broken packages:
-yarn patch-package
+npx patch-package
 
 # Fix Android dependency import statments:
 # Old native Android dependencies use outdated package names for their imports
@@ -34,13 +34,4 @@ node ./node_modules/.bin/rollup -c
 node -r sucrase/register ./scripts/stringifyBridge.ts
 
 # Create contract type definitions:
-yarn typechain
-
-# Bundle currency, swap, & rate plugins:
-core_assets="./android/app/src/main/assets/edge-core"
-if [ -d "$core_assets" ]; then
-  rm -r "$core_assets"
-fi
-mkdir -p "$core_assets"
-echo Webpacking plugins...
-node ./node_modules/.bin/webpack
+npm run typechain